Created chain for IO and CPU tasks. Added milliseconds to logger. Reduced clock frequency precision to prevent overflow

+++ scheduler.h
@@ -53,6 +53,6 @@ public:
 
 private:
     void process_interactive_tasks();
-    double getCurrentTicks();
+    long long getCurrentTicks();
     void addToCurrentTicks(std::chrono::milliseconds duration);
 };
